+++ framework/framework.go
@@ -1,7 +1,7 @@
 // Copyright 2022 Authors of spidernet-io
 // SPDX-License-Identifier: Apache-2.0
 
-package e2eframework
+package framework
 
 import (
 	"context"
