+++ framework/job.go
@@ -1,3 +1,3 @@
 // Copyright 2022 Authors of spidernet-io
 // SPDX-License-Identifier: Apache-2.0
-package e2eframework
+package framework
