+++ framework/kind.go
@@ -1,5 +1,5 @@
 // Copyright 2022 Authors of spidernet-io
 // SPDX-License-Identifier: Apache-2.0
-package e2eframework
+package framework
 
 // operate node container, like shutdown, ssh to login
