+++ framework/namespace.go
@@ -1,6 +1,6 @@
 // Copyright 2022 Authors of spidernet-io
 // SPDX-License-Identifier: Apache-2.0
-package e2eframework
+package framework
 
 import (
 	"fmt"
